feat(dagger.json): add new dagger.json file in golang directory
The new dagger.json file is added to the golang directory to specify the
configuration for the Dagger tool. This includes the name of the
project, the SDK being used, the source of the project, and the version
of the engine. This is necessary for Dagger to correctly build and run
the project.