feat(go.mod): add new dependencies to support additional features
The go.mod file has been updated with new dependencies. These
dependencies are required to support the new features that have been
added to the project. This includes libraries for cryptography, file
handling, network communication, and more. These libraries will help in
enhancing the functionality and performance of the application.