feat(README.md): add documentation for the new Kops Module
Documentation for the Kops Module has been added to the README.md to
provide users with information on managing Kubernetes clusters using
Kops. This includes details on the functions available within the
module, such as setting up containers with kubectl and Kops binaries,
exporting kubeconfig files, and configuring various aspects of the
Kubernetes cluster management.